+++ code/core/src/main/java/org/betonquest/betonquest/quest/condition/point/DefaultPointConditionFactory.java
@@ -1,0 +1,40 @@
+package org.betonquest.betonquest.quest.condition.point;
+
+import org.betonquest.betonquest.api.QuestException;
+import org.betonquest.betonquest.api.data.PointHolder;
+import org.betonquest.betonquest.api.instruction.Argument;
+import org.betonquest.betonquest.api.instruction.FlagArgument;
+import org.betonquest.betonquest.api.instruction.Instruction;
+import org.betonquest.betonquest.api.profile.Profile;
+import org.betonquest.betonquest.api.quest.TypeFactory;
+
+import java.util.function.Function;
+
+/**
+ * Factory to create {@link DefaultPointCondition}s for a given {@link PointHolder}.
+ */
+public class DefaultPointConditionFactory implements TypeFactory<DefaultPointCondition> {
+
+    /**
+     * Function to get the point holder from a profile.
+     */
+    private final Function<Profile, PointHolder> holderFunction;
+
+    /**
+     * Create a new factory for {@link PointHolder}.
+     *
+     * @param holderFunction the function to get the point holder from a profile
+     */
+    public DefaultPointConditionFactory(final Function<Profile, PointHolder> holderFunction) {
+        this.holderFunction = holderFunction;
+    }
+
+    @Override
+    public DefaultPointCondition parseInstruction(final Instruction instruction) throws QuestException {
+        final Argument<String> category = instruction.packageIdentifier().get();
+        final Argument<Number> count = instruction.number().get();
+        final FlagArgument<Boolean> equal = instruction.bool().getFlag("equal", true);
+        final FlagArgument<Number> fallback = instruction.number().getFlag("fallback", 0);
+        return new DefaultPointCondition(holderFunction, category, count, equal, fallback);
+    }
+}